1. Integrated gtest 2. Added a simple test-case to FileReader class as an example 3. Updated CMakeLists to compile gtest library as well as the unit-tests 4. Fixed a compiler issue on my gcc version 4.5.3 occuring in local_realignment.cpp

+++ openge/src/algorithms/local_realignment.cpp
@@ -782,7 +782,7 @@ void LocalRealignment::clean(ReadBin readsToClean) {
                     int neededBases = max(neededBasesToLeft, neededBasesToRight);
                     if ( neededBases > 0 ) {
                         int padLeft = max(leftmostIndex-neededBases, 1);
-                        int padRight = min(leftmostIndex+reference->size()+neededBases, (unsigned long)atol(referenceReader->getSequenceDictionary()[currentInterval->getContig()].Length.c_str()));
+                        int padRight = min((unsigned long)(leftmostIndex+reference->size()+neededBases), (unsigned long)atol(referenceReader->getSequenceDictionary()[currentInterval->getContig()].Length.c_str()));
                         reference = new string(referenceReader->getSubsequenceAt(currentInterval->getContig(), padLeft, padRight));
                         leftmostIndex = padLeft;
                     }
